+++ lib/twenv/command_mixins/archive_command.rb
@@ -25,18 +25,32 @@ module TWEnv::Command::ArchiveCommand
     write_archive path, read_archive(path).tap {|tweets| tweets.push(format_tweet(tweet)) }
   end
 
-  def resume_from_previous_archive(path)
-    tweet = read_archive(path)[-1]
-    self.max_id = tweet.id
-    line.puts "Continue from #{tweet.url} (#{tweet.created_at})"
-  end
-
   def read_archive(path)
     tweets = JSON.parse File.read(path)
     tweets.map {|tweet| TWEnv::Struct.from_hash(tweet)}
+  rescue Errno::ENOENT => ex
+    raise TWEnv::Error::NoSuchArchiveError, ex.message
   end
 
   def write_archive(path, tweets)
     File.write path, JSON.dump(tweets)
+  end
+
+  def resume_from_previous_archive(archive)
+    tweet = archive[-1]
+    if tweet
+      self.max_id = tweet.id
+      line.ok("Continue from #{tweet.url} (#{tweet.created_at})").end
+    else
+      line.warn("Archive is empty. Continue from start.").end
+    end
+  end
+
+  def finish_archive(path, old_archive, local_name:)
+    archive = read_archive(path)
+    return if old_archive == archive
+    sticky_locals.merge!(local_name => archive)
+    line.ok("Archive saved to #{relative_to_root(path)}").end
+    line.ok("Archive assigned to local variable `#{local_name}`").end
   end
 end