+++ lib/twenv/commands/archive_likes.rb
@@ -15,17 +15,22 @@ class TWEnv::ArchiveLikes < TWEnv::Command
   def process(user)
     self.user = user
     self.path = File.join storage_path, "#{user}.json"
-    opts['continue'] ? resume_from_previous_archive(path) : write_archive(path, [])
+    archive = opts['continue'] ? read_archive(path) : []
+    opts['continue'] ? resume_from_previous_archive(archive) : write_archive(path, archive)
     perform_action_on_tweets method(:read_tweets),
                              method(:archive_tweet),
                              method(:print_total),
-                             read_archive(path).map(&:id)
+                             archive.map(&:id)
+    line.end
+    finish_archive(path, archive, local_name: "archived_likes")
+  rescue TWEnv::Error::NoSuchArchiveError
+    line.error("Can't continue because an archive for #{user} doesn't exist").end
+  rescue Twitter::Error => ex
+    line.error("#{ex.message} (#{ex.class})").end
   rescue Interrupt
-    line.end_line
-  ensure
-    sticky_locals.merge!(archived_likes: read_archive(path))
-    line.puts "Archive saved to #{relative_to_root(path)}"
-    line.puts "Archive assigned to local variable `archived_likes`"
+    line.end
+    line.warn("Interrupt received").end
+    finish_archive(path, archive, local_name: "archived_likes")
   end
 
   def options(slop)
@@ -46,7 +51,7 @@ class TWEnv::ArchiveLikes < TWEnv::Command
   end
 
   def print_total(total)
-    line.rewind.print "#{total} likes archived"
+    line.rewind.ok "#{total} likes archived"
     throw(:cancel) if opts[:max].nonzero? && total == opts[:max]
   end