+++ lib/twenv/error.rb
@@ -1,0 +1,3 @@
+class TWEnv::Error < RuntimeError
+  NoSuchArchiveError = Class.new(self)
+end