+++ lib/twenv/line.rb
@@ -2,6 +2,9 @@
 
 class TWEnv::Line
   DESTRUCTIVE_BACKSPACE = "\b \b"
+  RED_ERROR_TEXT = Paint["ERR", "#CC0000", :bright]
+  GREEN_OK_TEXT  = Paint["OK", "#00FF00", :bright]
+  ORANGE_WARNING_TEXT = Paint["WARN", "#FFA500", :bright]
 
   def initialize(io)
     @io = io
@@ -9,7 +12,7 @@ class TWEnv::Line
   end
 
   def puts(str)
-    print(str).end_line
+    print(str).end
   end
 
   def print(str)
@@ -19,7 +22,22 @@ class TWEnv::Line
     self
   end
 
-  def end_line
+  def error(error_message)
+    print "#{RED_ERROR_TEXT} #{error_message}"
+    self
+  end
+
+  def ok(ok_message)
+    print "#{GREEN_OK_TEXT} #{ok_message}"
+    self
+  end
+
+  def warn(warn_message)
+    print "#{ORANGE_WARNING_TEXT} #{warn_message}"
+    self
+  end
+
+  def end
     @io.print "\n"
     self
   end